Nothing functional yet but I distrust my git skills rn so am committing before switching branches
